requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
bcrypt==4.1.2
psycopg2-binary==2.9.9
supabase==2.3.4
google-generativeai==0.3.2
Werkzeug==3.0.1
Jinja2==3.1.3
gunicorn==21.2.0